% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcodes.R
\name{BlockDecode}
\alias{BlockDecode}
\title{Block decoding of a code.}
\usage{
BlockDecode(code, block.encoder = NULL, visualize = FALSE)
}
\arguments{
\item{code}{The code to be decoded.}

\item{block.encoder}{Block encoder used for encoding.}

\item{visualize}{If TRUE a beamer PDF file is generated showing the decode process.}
}
\value{
The decoded message as binary Vector
}
\description{
Decodes a block codeword.
}
\examples{
coder <- BlockGenerateEncoderBCH(15,3)
plain <- c(1,0,0,1,1)

# standard encoding and decoding
coded <- BlockEncode(plain, coder)
BlockDecode(coded, coder)

# with visualization
# coded <- BlockEncode(plain, coder)
# BlockDecode(coded, coder, visualize = TRUE)

# with message distortion
coded <- BlockEncode(plain, coder)
noisy <- ApplyNoise(coded, SNR.db = 3, binary = TRUE)
BlockDecode(noisy, coder)
}
\author{
Benedikt Wimmer
}

