% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcodes.R
\name{BlockEncode}
\alias{BlockEncode}
\title{Block encoding of a message.}
\usage{
BlockEncode(message, block.encoder = NULL, visualize = FALSE)
}
\arguments{
\item{message}{The message to be encoded.}

\item{block.encoder}{Block encoder used for encoding.}

\item{visualize}{If TRUE a beamer PDF file is generated showing the encode process.}
}
\value{
The encoded message as binary vector.
}
\description{
Produces a block code of a message.
}
\examples{
coder <- BlockGenerateEncoderBCH(15,3)
plain <- c(1,0,0,1,1)

# standard encoding
BlockEncode(plain, coder)

# standard encoding with visualization
BlockEncode(plain, coder, visualize = TRUE)

# use default values
BlockEncode(plain)
}
\author{
Benedikt Wimmer
}

