% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcodes.R
\name{BlockGenerateEncoderBCH}
\alias{BlockGenerateEncoderBCH}
\title{Generate BCH encoder.}
\usage{
BlockGenerateEncoderBCH(code.length = 15, code.t = 3)
}
\arguments{
\item{code.length}{Length of the encoded message}

\item{code.t}{Error correcting capability}
}
\value{
A block encoder represented as a list containing:
    type, code.length, code.m(minimal m for 2^m >= code.length), data.length, code.t,
    gen.poly,
    2 Vectors alpha_to and index_of describing the Galois Field GF(2^m)
}
\description{
Generates a block encoder for BCH block codes.
}
\details{
Apart from necessary parameters like data.length, a BCH encoder consists of a Generator Polynomial, which is needed for encoding,
and the polynomial and index representation of the underlying Galois Field GF(code.length+1), which is needed for constructing and solving the
error location Polynomial.
}
\examples{
# standard bch encoder with code-rate = 0.33
BlockGenerateEncoderBCH(15,3)
}
\author{
Benedikt Wimmer
}

