% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcodes.R
\name{BlockGenerateEncoderHamming}
\alias{BlockGenerateEncoderHamming}
\title{Generate Hamming encoder.}
\usage{
BlockGenerateEncoderHamming(code.length = 7, data.length = 4)
}
\arguments{
\item{code.length}{Length of the encoded message}

\item{data.length}{Length of the plain message}
}
\value{
A block encoder represented as a list containing:
    type, code.length, data.length,
    2 Matrices gen.matrix and check.matrix, being the Generator Matrix and the Parity Check Matrix
}
\description{
Generates a block encoder for Hamming block codes.
}
\details{
The resulting Hamming Code is obtained by constructing a (code.length - data.length) x code.length Parity-Check-Matrix.
The columns are the binary vectors of the integers 1 to code.length. Column Permutations are then applied to get the
Form (t(A) | Ir). The resulting Generator Matrix is (In | A).
}
\examples{
# standard Hamming encoder with code-rate = 0.57
BlockGenerateEncoderHamming(7,4)
}
\author{
Benedikt Wimmer
}

