% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcodes.R
\name{BlockOpenPDF}
\alias{BlockOpenPDF}
\title{Open visualization PDF.}
\usage{
BlockOpenPDF(encode = TRUE, hamming = FALSE, simulation = FALSE)
}
\arguments{
\item{encode}{Flag to open encode PDFs (if true) or decode PDFs (if false).}

\item{hamming}{Flag to open encode or decode PDFs for Hamming codes. For BCH use false.}

\item{simulation}{Flag to open simulation PDFs. This flag has highest precedence
meaning that if the simulation flag is TRUE the function will look for the
simulation PDF. Only if FALSE the others are evaluated.}
}
\description{
With this function it is easy to reopen the PDF files created with
    \code{\link{BlockEncode}}, \code{\link{BlockDecode}},
    and \code{\link{BlockSimulation}}.
    The files are stored in the program files of R.
    If the corresponding file does not exist yet an error message is printed.
}
\examples{
# open encode for Hamming PDF
BlockOpenPDF(encode = TRUE, hamming = TRUE)

# open decode for BCH PDF
BlockOpenPDF(encode = FALSE, hamming = FALSE)

# open block simulation PDF
BlockOpenPDF(simulation = TRUE)
}
\author{
Benedikt Wimmer
}

