% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ChannelcodingSimulation}
\alias{ChannelcodingSimulation}
\title{Channelcoding Simulation}
\usage{
ChannelcodingSimulation(msg.length = 100, min.db = 0.1, max.db = 2,
  db.interval = 0.1, iterations.per.db = 100, turbo.decode.iterations = 5,
  visualize = FALSE)
}
\arguments{
\item{msg.length}{Message length of the randomly created messages to be encoded.}

\item{min.db}{Minimum SNR to be tested.}

\item{max.db}{Maximum SNR to be tested.}

\item{db.interval}{Step between two SNRs tested.}

\item{iterations.per.db}{Amount of iterations each signal/noise ration step.}

\item{turbo.decode.iterations}{Amount of decoding iterations inside the turbo decoder.}

\item{visualize}{If true a PDF report is generated.}
}
\value{
Dataframe containing the bit-error-rates for each coding technique
    and each SNR tested.
}
\description{
Simulation of channelcoding techniques (blockcodes, convolutional codes
    and turbo codes) and comparison of their bit-error-rates.
}

