% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{conv_decode}
\alias{conv_decode}
\title{convolutional decoding of a code (soft decision)}
\usage{
conv_decode(code, convEncoder, terminate = TRUE)
}
\arguments{
\item{code}{the code to be decoded}

\item{convEncoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code was terminated, default: TRUE}
}
\value{
the decoded message, list(softOutput, hardOutput)
}
\description{
\code{conv_decode} decodes a codeword that was encoded with the given encoder.
This decoder is a soft-input soft-output decoder
}
\author{
Martin Nocker
}

