% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{conv_decode_hard}
\alias{conv_decode_hard}
\title{convolutional decoding of a code (hard decision)}
\usage{
conv_decode_hard(code, convEncoder, terminate = TRUE)
}
\arguments{
\item{code}{the code to be decoded}

\item{convEncoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code was terminated, default: TRUE}
}
\value{
the hard-decoded message vector
}
\description{
\code{conv_decode} decodes a codeword that was encoded with the given encoder.
This decoder is a hard-decision decoder
}
\author{
Martin Nocker
}

