% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{conv_encode}
\alias{conv_encode}
\title{convolutional encoding of a message}
\usage{
conv_encode(message, convEncoder, terminate = TRUE)
}
\arguments{
\item{message}{the message to be encoded}

\item{convEncoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code should be terminated, default: TRUE}
}
\value{
the encoded message
}
\description{
\code{conv_encode} produces a convolutional code of a message based on the encoder passed as an argument
}
\author{
Martin Nocker
}

