% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{ConvDecodeHard}
\alias{ConvDecodeHard}
\title{convolutional decoding of a code (hard decision)}
\usage{
ConvDecodeHard(code, conv.encoder, terminate = TRUE)
}
\arguments{
\item{code}{the code to be decoded}

\item{conv.encoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code was terminated, default: TRUE}
}
\value{
the hard-decoded message vector
}
\description{
\code{ConvDecodeHard} decodes a codeword that was encoded with the given
encoder. This decoder is a hard-decision decoder
}
\examples{
nsc <- GenerateNscEncoder(2,2,c(7,5))
coded <- ConvEncode(c(1,0,0,1,1), nsc)
ConvDecodeHard(coded, nsc)
}
\author{
Martin Nocker
}

