% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvDecodeSoft}
\alias{ConvDecodeSoft}
\title{Convolutional decoding of a code (soft decision).}
\usage{
ConvDecodeSoft(code, conv.encoder = NULL, terminate = TRUE,
  punctuation.matrix = NULL, visualize = FALSE)
}
\arguments{
\item{code}{The code to be decoded.}

\item{conv.encoder}{Convolutional encoder used for encoding.}

\item{terminate}{flag If the code was terminated.}

\item{punctuation.matrix}{If not null the code is depunctured prior to the decode algorithm.}

\item{visualize}{If TRUE a beamer PDF file is generated showing the decode process.}
}
\value{
The decoded message, list(softOutput, hardOutput)
}
\description{
Decodes a convolutional codeword.
This decoder is a soft-input soft-output decoder.
}
\examples{
coder <- ConvGenerateEncoder(2,2,c(7,5))
coded <- ConvEncode(c(1,0,0,1,1), coder)
ConvDecodeSoft(coded, coder)
}
\author{
Martin Nocker
}

