% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{ConvEncode}
\alias{ConvEncode}
\title{convolutional encoding of a message}
\usage{
ConvEncode(message, conv.encoder, terminate = TRUE)
}
\arguments{
\item{message}{the message to be encoded}

\item{conv.encoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code should be terminated, default: TRUE}
}
\value{
the encoded message
}
\description{
\code{ConvEncode} produces a convolutional code of a message based on the encoder passed as an argument
}
\author{
Martin Nocker
}

