% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvGetPunctuationMatrix}
\alias{ConvGetPunctuationMatrix}
\title{Get punctuation matrix from vector.}
\usage{
ConvGetPunctuationMatrix(punctuation.vector, conv.coder)
}
\arguments{
\item{punctuation.vector}{Vector containing the punctuation information which will
be transformed to a punctuation matrix.}

\item{conv.coder}{Convolutional coder which is used for the matrix dimensions.}
}
\value{
Punctuation matrix suitable for \code{\link{ConvEncode}}, \code{\link{ConvDecodeSoft}},
    \code{\link{ConvDecodeHard}} and \code{\link{ConvSimulation}}.
}
\description{
Creates a punctuation matrix from the passed punctuation vector and the
passed coder.
}

