% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvolutionOpenPDF}
\alias{ConvolutionOpenPDF}
\title{Open visualization PDF}
\usage{
ConvolutionOpenPDF(encode = TRUE, punctured = FALSE, simulation = FALSE)
}
\description{
Opens a PDF of the encode, decode or simulation visualizations or prints a message
if the corresponding file does not exist yet.
}

