% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvolutionSimulation}
\alias{ConvolutionSimulation}
\title{Convolutional Simulation}
\usage{
ConvolutionSimulation(coder = NULL, msg.length = 100,
  iterations.per.db = 100, min.db = 0.1, max.db = 2, db.interval = 0.1,
  punctuation.matrix = NULL)
}
\arguments{
\item{coder}{convolutional coder used for the simulation}

\item{msg.length}{message length of the encoded messages}

\item{iterations.per.db}{number of encode and decode processes per SNR}

\item{min.db}{minimum SNR to be tested}

\item{max.db}{maximum SNR to be tested}

\item{db.interval}{step between two SNRs tested}

\item{punctuation.matrix}{if not null the process involves the punctuation}
}
\value{
dataframe containing the bit-error-rates for several SNRs
}
\description{
Simulation of a convolutional encode and decode process over a noisy channel
}
\examples{
coder <- GenerateConvEncoder(2,2,c(7,5))
ConvolutionSimulation(coder, 100, 100, 0.1, 2.0, 0.1, NULL)
}
\author{
Martin Nocker
}

