% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvOpenPDF}
\alias{ConvOpenPDF}
\title{Open visualization PDF.}
\usage{
ConvOpenPDF(encode = TRUE, punctured = FALSE, simulation = FALSE)
}
\arguments{
\item{encode}{Flag to open encode PDFs (if true) or decode PDFs (if false).}

\item{punctured}{Flag to open encode or decode PDFs with punctuation.}

\item{simulation}{Flag to open simulation PDFs. This flag has highest precedence
meaning that if the simulation flag is TRUE the function will look for the
simulation PDF. Only if FALSE the others are evaluated.}
}
\description{
With this function it is easy to reopen the PDF files created with
    \code{\link{ConvEncode}}, \code{\link{ConvDecodeSoft}},
    \code{\link{ConvDecodeHard}} and \code{\link{ConvSimulation}}.
    The files are stored in the program files of R.
    If the corresponding file does not exist yet an error message is printed.
}
\examples{
# open encode without punctuation PDF
ConvOpenPDF(encode = TRUE, punctured = FALSE)

# open decode with punctuation PDF
ConvOpenPDF(encode = FALSE, punctured = TRUE)

# open convolution simulation PDF
ConvOpenPDF(simulation = TRUE)
}
\author{
Martin Nocker
}

