% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{decode}
\alias{decode}
\title{general decode function}
\usage{
decode(msg, type, params, visualize)
}
\arguments{
\item{msg}{message to decode}

\item{type}{the type of code you want to use}

\item{params}{A vector with arguments}

\item{visualize}{A flag for enabling visualization}
}
\value{
decoded message
}
\description{
Is used as a generic decode funtion that will forward the call depending on the type
}
\author{
Bene Wimmer
}

