% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{GenerateConvEncoder}
\alias{GenerateConvEncoder}
\title{generate convolutional encoder}
\usage{
GenerateConvEncoder(N, M, generators)
}
\arguments{
\item{N}{numer ob output symbols per input symbol}

\item{M}{memory length of the encoder}

\item{generators}{vector of generator polynoms (one for each output symbol)}
}
\value{
a convolutional encoder represented as a list containing:
N, M, 4 matrices: nextState, previousState, output, termination, nsc (flag)
}
\description{
generates a convolutional encoder for nonrecursive convolutional codes
}
\examples{
GenerateConvEncoder(2,2,c(7,5))
}
\author{
Martin Nocker
}

