% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{generateConvEncoder_nsc}
\alias{generateConvEncoder_nsc}
\title{generate nsc encoder}
\usage{
generateConvEncoder_nsc(N, M, generators)
}
\arguments{
\item{N}{numer ob output symbols per input symbol}

\item{M}{constraint length (memory length of the encoder)}

\item{generators}{vector of generator polynoms (one for each output symbol)}
}
\value{
a convolutional encoder represented as a list containing:
N, M, 4 matrices: nextState, previousState, output, termination, nsc (flag)
}
\description{
generates a convolutional encoder for nonsystematic convolutional codes (nsc)
}
\author{
Martin Nocker
}

