% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{GetPunctuationMatrix}
\alias{GetPunctuationMatrix}
\title{get punctuation matrix from vector}
\usage{
GetPunctuationMatrix(punctuation.vector, coder.info)
}
\arguments{
\item{punctuation.vector}{vector containing the punctuation information which will
be transformed to a punctuation matrix}

\item{coder.info}{channelcoder which is used for the matrix dimensions}
}
\value{
punctuation matrix suitable for encode and decode
}
\description{
creates a punctuation matrix from the passed punctuation vector and the
passed coder
}

