% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{testBCH}
\alias{testBCH}
\title{Testing BCH codes}
\usage{
testBCH(data, length, t, visualize = FALSE)
}
\arguments{
\item{data}{MSG}

\item{length}{Length of MSG}

\item{t}{Errors to correct}
}
\value{
String
}
\description{
Testing BCH codes
}
\author{
Bene Wimmer
}

