% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_code_rcpp.R
\name{TurboDecode}
\alias{TurboDecode}
\title{Turbo Decode}
\usage{
TurboDecode(message, permutation, iterations, encoder.info, parity.index)
}
\arguments{
\item{message}{Nachricht die dekodiert werden sollte}

\item{permutation}{Permutationsvektor der mittels \code{\link{TurboGetPermutation}} erzeugt werden sollte}

\item{encoder.info}{Dekodierer}

\item{parity.index}{Index des zu verwendenten Ausgangs des Kodieres}
}
\value{
Deodierte Nachricht
}
\description{
Dekodiert eine Nachricht mittels dem Turbo-Code-Verfahren. Dabei wird die Nachricht in 2 systematische Dekodierer
gesteckt, wobei bei einem Kodierer die Nachricht permutiert verarbeitet wird. Der mitgegebene Kodierer muss
mindestens 2 Ausgänge haben und sollte Systematisch sein. Dadurch wird der erste Ausgang automatisch
bei beiden Deodierern durchgeschalten. Nachdem die Nachricht durche beide Dekodierer durch ist, kann dieser Schritt
mehrmals wiederholt werden. Mit dem Parameter iterations, kann die Anzahl der Durchläufe verändert werden.
Je mehr Durchläufe, desto besser das Ergebnis. Das Prinzip der Turbo-Codes ist, dass die Dekodierung mit Soft-Werten
vollzogen wird. Das heißt, dass beim Eingang der tatsächliche Signalpegel berücksichtigt wird und beim Ausgang ein
Wahrscheinlichkeitswert berechnet wird, der angibt, wie wahrscheinlich ein Bit am Ausgang ist
}
\author{
Witsch Daniel
}

