% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_code_rcpp.R
\name{TurboEncode}
\alias{TurboEncode}
\title{Turbo Encode}
\usage{
TurboEncode(message, permutation, encoder.info, parity.index = encoder.info$N)
}
\arguments{
\item{message}{Nachricht die kodiert werden sollte}

\item{permutation}{Permutationsvektor der mittels \code{\link{TurboGetPermutation}} erzeugt werden sollte}

\item{encoder.info}{Kodierer}

\item{parity.index}{Index des zu verwendenten Ausgangs des Kodieres}
}
\value{
Kodierte Nachricht, die aus Ausgangsnachricht und den beiden kodierten Nachrichten besteht
}
\description{
Kodiert eine Nachricht mittels dem Turbo-Code-Verfahren. Dabei wird die Nachricht in 2 systematische Kodierer
gesteckt, wobei bei einem Kodierer die Nachricht permutiert verarbeitet wird. Der mitgegebene Kodierer muss
mindestens 2 Ausgänge haben und sollte Systematisch sein. Dadurch wird der erste Ausgang automatisch
bei beiden Kodierern durchgeschalten, somit muss die Ausgangsnachricht nur einmal übertragen werden,
dadurch wird die Koderate verbessert. Um den Minimalabstand der Kodierer zu erhöhen, wird die Nachricht
permutiert in den 2ten Kodierer geschickt. Am Ende wird die Ausgangsnachricht mit den beiden kodierten Nachrichten
aus den Kodierern verknüpft und retour gegeben. Da nur ein Ausgang von einem Kodierer verwendet wir, muss der Index
des Ausgangs angegeben werden.
}
\author{
Witsch Daniel
}

