% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo_code_rcpp.R
\name{TurboGetPunctuationMatrix}
\alias{TurboGetPunctuationMatrix}
\title{Create a punctuation matrix from a vector.}
\usage{
TurboGetPunctuationMatrix(punctuation.vector, visualize = FALSE)
}
\arguments{
\item{punctuation.vector}{Vector which will be mapped to the matrix.}

\item{visualize}{Flag to visualize the resulting punctuation matrix.}
}
\value{
Punctuation matrix for encoder and decoder.
}
\description{
This is a helper function to create a correct punctuation matrix. The punctuation.vector
should only contain 0 and 1 and has to be a length which is a multiple of 3. The resulting
punctuation matrix should not contain a 0 column because this could cause an error during
decoding.
}
\examples{
TurboGetPunctuationMatrix(c(1, 1, 0, 1, 0, 1), TRUE)

}

