% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{applyNoise}
\alias{applyNoise}
\title{Failure function}
\usage{
applyNoise(msg, SNR.db = 5, visualize = FALSE)
}
\arguments{
\item{msg}{message to alter}

\item{visualize}{A flag for enabling visualization}

\item{params}{A vector with arguments}
}
\value{
message with errors
}
\description{
This function will randomly disturb a clean message
}
\author{
Bene Wimmer
}

