% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ApplyNoise}
\alias{ApplyNoise}
\title{Message distortion.}
\usage{
ApplyNoise(msg, SNR.db = 3, binary = FALSE)
}
\arguments{
\item{msg}{Message to alter.}

\item{SNR.db}{Signal-Noise-Ratio in dB simulating the noisy channel.}

\item{binary}{False = Soft Output, True = Hard Output}
}
\value{
Distorted message containing noise.
}
\description{
This function will randomly distort a clean message.
}
\author{
Bene Wimmer
}

