% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcodes.R
\name{BlockGenerateEncoderHamming}
\alias{BlockGenerateEncoderHamming}
\title{Generate Hamming encoder.}
\usage{
BlockGenerateEncoderHamming(code.length = 7, data.length = 4)
}
\arguments{
\item{code.length}{Length of the encoded message}

\item{data.length}{Length of the plain message}
}
\value{
A block encoder represented as a list containing:
    type, code.length, data.length,
    2 Matrices gen.matrix and check.matrix, being the Generator Matrix and the Parity Check Matrix
}
\description{
Generates a block encoder for Hamming block codes.
}
\details{
To come.
}
\examples{
# standard Hamming encoder with code-rate = 0.57
BlockGenerateEncoderHamming(7,4)
}
\author{
Benedikt Wimmer
}

