% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelcoding.R
\docType{package}
\name{channelcoding}
\alias{channelcoding}
\alias{channelcoding-package}
\title{Channelcoding Functions for R}
\description{
The channelcoding package provides functionality for three different
channelcodes: block codes, convolutional codes and turbo codes.
Self-defined encoders, encoding, decoding, message distortion as well
as simulations and visualizations are supported.
}
\author{
Martin Nocker, Benedikt Wimmer, Daniel Witsch
}

