% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvDecode}
\alias{ConvDecode}
\title{convolutional decoding of a code (soft decision)}
\usage{
ConvDecode(code, conv.encoder, terminate = TRUE, puncturing.matrix = NULL)
}
\arguments{
\item{code}{the code to be decoded}

\item{conv.encoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code was terminated, default: TRUE}
}
\value{
the decoded message, list(softOutput, hardOutput)
}
\description{
\code{ConvDecode} decodes a convolutional codeword
This decoder is a soft-input soft-output decoder
}
\examples{
coder <- GenerateConvEncoder(2,2,c(7,5))
coded <- ConvEncode(c(1,0,0,1,1), coder)
ConvDecode(coded, coder)
}
\author{
Martin Nocker
}

