% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvDecodeHard}
\alias{ConvDecodeHard}
\title{convolutional decoding of a code (hard decision)}
\usage{
ConvDecodeHard(code, conv.encoder = NULL, terminate = TRUE,
  punctuation.matrix = NULL, visualize = FALSE)
}
\arguments{
\item{code}{the code to be decoded}

\item{conv.encoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code was terminated, default: TRUE}

\item{punctuation.matrix}{if not null the code is depunctured prior to the decode algorithm}

\item{visualize}{if TRUE a beamer PDF file is generated showing the decode process}
}
\value{
the hard-decoded message vector
}
\description{
\code{ConvDecodeHard} decodes a codeword that was encoded with the given
encoder. This decoder is a hard-decision decoder
}
\examples{
coder <- GenerateConvEncoder(2,2,c(7,5))
coded <- ConvEncode(c(1,0,0,1,1), coder)
ConvDecodeHard(coded, coder)
}
\author{
Martin Nocker
}

