% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvDecodeSoft}
\alias{ConvDecodeSoft}
\title{Convolutional decoding of a code (soft decision).}
\usage{
ConvDecodeSoft(code, conv.encoder = NULL, terminate = TRUE,
  punctuation.matrix = NULL, visualize = FALSE)
}
\arguments{
\item{code}{The code to be decoded.}

\item{conv.encoder}{Convolutional encoder used for encoding.}

\item{terminate}{flag If the code is terminated.}

\item{punctuation.matrix}{If not null the code is depunctured prior to the decode algorithm.}

\item{visualize}{If TRUE a beamer PDF file is generated showing the decode process.}
}
\value{
The decoded message, list(softOutput, hardOutput)
}
\description{
Decodes a convolutional codeword using soft decision decoding.
}
\examples{
coder <- ConvGenerateEncoder(2,2,c(7,5))
plain <- c(1,0,0,1,1)

# standard encoding and decoding with termination
coded <- ConvEncode(plain, coder)
ConvDecodeSoft(coded, coder)

# without termination
coded <- ConvEncode(plain, coder, terminate = FALSE)
ConvDecodeSoft(coded, coder, terminate = FALSE)

# with punctuation and visualization
# punctuation <- ConvGetPunctuationMatrix(c(1,1,0,1,1,0), coder)
# coded <- ConvEncode(plain, coder, punctuation.matrix = punctuation)
# ConvDecodeSoft(coded, coder, punctuation.matrix = punctuation, visualize = TRUE)

# with message distortion
coded <- ConvEncode(plain, coder)
noisy <- ApplyNoise(coded, SNR.db = 3)
ConvDecodeSoft(noisy, coder)
}
\author{
Martin Nocker
}

