% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvEncode}
\alias{ConvEncode}
\title{convolutional encoding of a message}
\usage{
ConvEncode(message, conv.encoder, terminate = TRUE)
}
\arguments{
\item{message}{the message to be encoded}

\item{conv.encoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code should be terminated, default: TRUE}
}
\value{
the encoded message
}
\description{
\code{ConvEncode} produces a convolutional code of a message
}
\examples{
coder <- GenerateConvEncoder(2,2,c(7,5))
ConvEncode(c(1,0,0,1,1), coder)
}
\author{
Martin Nocker
}

