% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvEncode}
\alias{ConvEncode}
\title{Convolutional encoding of a message.}
\usage{
ConvEncode(message, conv.encoder = NULL, terminate = TRUE,
  punctuation.matrix = NULL, visualize = FALSE)
}
\arguments{
\item{message}{The message to be encoded.}

\item{conv.encoder}{Convolutional encoder used for encoding.}

\item{terminate}{Flag if the code should be terminated.}

\item{punctuation.matrix}{If not null the encoded message is punctured with the punctuation matrix.}

\item{visualize}{If TRUE a beamer PDF file is generated showing the encode process.}
}
\value{
The encoded message with signal values {-1,+1} which map to {1,0} respectively.
}
\description{
Produces a convolutional code of a message.
}
\examples{
coder <- ConvGenerateEncoder(2,2,c(7,5))
plain <- c(1,0,0,1,1)

# standard encoding with termination
ConvEncode(plain, coder)

# without termination
ConvEncode(plain, coder, terminate = FALSE)

# with punctuation and visualization
# punctuation <- ConvGetPunctuationMatrix(c(1,1,0,1,1,0), coder)
# ConvEncode(plain, coder, punctuation.matrix = punctuation, visualize = TRUE)

# use default values
ConvEncode(plain)
}
\author{
Martin Nocker
}

