% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvEncode}
\alias{ConvEncode}
\title{convolutional encoding of a message}
\usage{
ConvEncode(message, conv.encoder = NULL, terminate = TRUE,
  punctuation.matrix = NULL, visualize = FALSE)
}
\arguments{
\item{message}{the message to be encoded}

\item{conv.encoder}{convolutional encoder used for encoding [list]}

\item{terminate}{flag if the code should be terminated, default: TRUE}

\item{punctuation.matrix}{if not null the encoded message is punctured with the punctuation matrix}

\item{visualize}{if TRUE a beamer PDF file is generated showing the encode process}
}
\value{
the encoded message
}
\description{
\code{ConvEncode} produces a convolutional code of a message
}
\examples{
coder <- GenerateConvEncoder(2,2,c(7,5))
ConvEncode(c(1,0,0,1,1), coder)
}
\author{
Martin Nocker
}

