% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvSimulation}
\alias{ConvSimulation}
\title{Convolutional Simulation}
\usage{
ConvSimulation(coder = NULL, msg.length = 100, iterations.per.db = 100,
  min.db = 0.1, max.db = 2, db.interval = 0.1,
  punctuation.matrix = NULL, visualize = TRUE)
}
\arguments{
\item{coder}{convolutional coder used for the simulation. Can be created via
\code{\link{GenerateConvEncoder}} or \code{\link{GenerateRscEncoder}}.}

\item{msg.length}{message length of the randomly created messages to be encoded}

\item{iterations.per.db}{number of encode and decode processes per SNR}

\item{min.db}{minimum SNR to be tested}

\item{max.db}{maximum SNR to be tested}

\item{db.interval}{step between two SNRs tested}

\item{punctuation.matrix}{if not null the process involves the punctuation. Can
be created via \code{\link{GetPunctuationMatrix}}.}

\item{visualize}{if true a PDF report is generated}
}
\value{
dataframe containing the bit-error-rates for each SNR tested
}
\description{
Simulation of a convolutional encode and decode process over a noisy channel
}
\examples{
coder <- GenerateConvEncoder(2,2,c(7,5))
ConvSimulation(coder, 100, 100, 0.1, 2.0, 0.1, NULL)

#all default parameters
ConvSimulation()

#without punctuation
coder <- GenerateConvEncoder(2,2,c(7,5))
ConvSimulation(coder, 10, 50, 0.01, 1, 0.05, NULL, TRUE)
}
\author{
Martin Nocker
}

