% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{ConvSimulation}
\alias{ConvSimulation}
\title{Convolutional Simulation.}
\usage{
ConvSimulation(conv.coder = NULL, msg.length = 100,
  iterations.per.db = 100, min.db = 0.1, max.db = 2, db.interval = 0.1,
  punctuation.matrix = NULL, visualize = FALSE)
}
\arguments{
\item{msg.length}{Message length of the randomly created messages to be encoded.}

\item{iterations.per.db}{Number of encode and decode processes per SNR.}

\item{min.db}{Minimum SNR to be tested.}

\item{max.db}{Maximum SNR to be tested.}

\item{db.interval}{Step between two SNRs tested.}

\item{punctuation.matrix}{If not null the process involves the punctuation. Can
be created via \code{\link{ConvGetPunctuationMatrix}}.}

\item{visualize}{If true a PDF report is generated.}

\item{coder}{Convolutional coder used for the simulation. Can be created via
\code{\link{ConvGenerateEncoder}} or \code{\link{ConvGenerateRscEncoder}}.}
}
\value{
Dataframe containing the bit-error-rates for each SNR tested.
}
\description{
Simulation of a convolutional encode and decode process over a noisy channel.
}
\examples{
#all default parameters
ConvSimulation()

#without punctuation
coder <- ConvGenerateEncoder(2,2,c(7,5))
ConvSimulation(coder, 10, 50, 0.01, 1, 0.05, NULL, FALSE)
}
\author{
Martin Nocker
}

