% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelcoding.R
\name{encode}
\alias{encode}
\title{general encdoe funtion}
\usage{
encode(msg, type, params, visualize)
}
\arguments{
\item{msg}{message to encode}
}
\value{
encoded message
}
\description{
general encdoe funtion
}
\author{
Witsch Daniel
}

