% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interface.R
\name{encode}
\alias{encode}
\title{generic encode function}
\usage{
encode(msg, type, params, visualize)
}
\arguments{
\item{msg}{message to encode}

\item{type}{the type of code you want to use}

\item{params}{A vector with arguments}

\item{visualize}{A flag for enabling visualization}
}
\value{
encoded message
}
\description{
Is used as a generic encode funtion that will forward the call depending on the type
}
\author{
Bene Wimmer
}

