% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_rcpp.R
\name{GenerateRscEncoder}
\alias{GenerateRscEncoder}
\title{generate rsc encoder}
\usage{
GenerateRscEncoder(N, M, generators)
}
\arguments{
\item{N}{numer ob output symbols per input symbol}

\item{M}{constraint length (memory length of the encoder)}

\item{generators}{vector of generator polynoms (one for each output symbol and one for the recursion)}
}
\value{
a convolutional encoder represented as a list containing:
N, M, 4 matrices: nextState, previousState, output, termination, nsc (flag)
}
\description{
generates a convolutional encoder for recursive systematic codes (rsc)
}
\author{
Martin Nocker
}

