% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PlotSimulationData}
\alias{PlotSimulationData}
\title{Show simulation data in a plot}
\usage{
PlotSimulationData(...)
}
\arguments{
\item{...}{Dataframes created by the simulation functions.}
}
\description{
Shows passed simulation data (can be created by the functions
\code{\link{ConvSimulation}}, \code{\link{TurboSimulation}}
and <BlockSimulation> containing the bit-error-rates for several SNRs
in one plot for easy comparison.
}

