% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/turbo_code_rcpp.R
\name{TurboOpenPDF}
\alias{TurboOpenPDF}
\title{Open the beforehand created visualization PDF files again.}
\usage{
TurboOpenPDF(encode = TRUE, punctured = FALSE, simulation = FALSE)
}
\arguments{
\item{encode}{flag to open the encode pdfs}

\item{punctured}{flag to open the decode pdfs}

\item{simulation}{flag to open the simulation pdf}
}
\description{
With this function it is easy to reopen the PDF files which will be created with
\code{\link{TurboEncode}} and \code{\link{TurboDecode}}. The files are stored in the
program files of R. (example path: "C:\Program Files\R\R-3.2.4\library\channelcoding\pdf")
}

